refactor: sentry middleware
